Align PONIX home preview with public rendering
Extract the home overview assembly out of the public page so the CMS canvas can render the same HomeSection instead of the generic section-card fallback. Draft home composition now resolves to a typed home preview while keeping the existing public page behavior.

Constraint: PONIX live composition must match the public page renderer for CMS trust.

Rejected: Keep the generic preview for home | it diverges from the live page and hides visual regressions.

Confidence: high

Scope-risk: moderate

Directive: Future page preview work should add typed preview branches instead of accepting the generic fallback for public pages.

Tested: pnpm exec tsc --noEmit; pnpm run lint; pnpm run build; local public home smoke via localhost:3000 rendered HomeSection content and no browser errors.

Not-tested: Authenticated local PONIX canvas visual smoke because local canvas redirects to sign-in without an admin session.
